Feat[shortcuts]: move shortcut pinner to Instances & disable them on delete
