+++ app_pojavlauncher/src/main/java/net/kdt/pojavlaunch/instances/Instance.java
@@ -114,7 +114,7 @@ public class Instance extends DisplayInstance {
         if(sharedData) return Instances.SHARED_DATA_DIRECTORY;
         return mInstanceRoot;
     }
-    public File getInstanceRoot(){
+    File getInstanceRoot(){
         return mInstanceRoot;
     }
 }
